Create a "trait" to label the Colony Leader.
This is so that the leadership can be seen in the Colonist's Pin & Data
screens, and so that they can be searched for in the Colonist overview
screens. The display_name of the trait is set to the Leader's Title, and
the description to the Sponsor + Leader Title.
Example: "Church of the New Ark Oracle".